+++ examples/simple/settings/bootstrap3_theme.py
@@ -3,7 +3,6 @@ from .base import *
 INSTALLED_APPS = list(INSTALLED_APPS)
 
 try:
-    INSTALLED_APPS.remove('south') if 'south' in INSTALLED_APPS else None
     INSTALLED_APPS.remove('tinymce') if 'tinymce' in INSTALLED_APPS else None
 except Exception as err:
     pass
@@ -17,16 +16,3 @@ try:
         if 'admin_tools.dashboard' in INSTALLED_APPS else None
 except Exception as err:
     pass
-
-FOBI_DEFAULT_THEME = 'foundation5'
-
-INSTALLED_APPS = list(INSTALLED_APPS)
-
-INSTALLED_APPS.append(
-    'fobi.contrib.themes.foundation5.widgets.form_elements.'
-    'date_foundation5_widget'
-)
-INSTALLED_APPS.append(
-    'fobi.contrib.themes.foundation5.widgets.form_elements.'
-    'datetime_foundation5_widget'
-)