+++ examples/simple/settings/bootstrap3_theme_feincms.py
@@ -1,3 +1,4 @@
+from nine.versions import DJANGO_GTE_1_8
 from .base import *
 
 INSTALLED_APPS = list(INSTALLED_APPS)
@@ -7,15 +8,11 @@ INSTALLED_APPS += [
     'fobi.contrib.apps.feincms_integration',  # Fobi FeinCMS app
 
     'page',  # Example
+
+    'tinymce',  # TinyMCE
 ]
 
 try:
-    # INSTALLED_APPS.remove('admin_tools') \
-    #     if 'admin_tools' in INSTALLED_APPS \
-    #     else None
-    # INSTALLED_APPS.remove('admin_tools.menu') \
-    #     if 'admin_tools.menu' in INSTALLED_APPS \
-    #     else None
     INSTALLED_APPS.remove('admin_tools.dashboard') \
         if 'admin_tools.dashboard' in INSTALLED_APPS else None
 except Exception as err:
@@ -24,3 +21,10 @@ except Exception as err:
 FEINCMS_RICHTEXT_INIT_CONTEXT = {
     'TINYMCE_JS_URL': STATIC_URL + 'tiny_mce/tiny_mce.js',
 }
+
+if DJANGO_GTE_1_8:
+
+    MIGRATION_MODULES = {
+        'fobi': 'fobi.migrations',
+        'db_store': 'fobi.contrib.plugins.form_handlers.db_store.migrations',
+    }